pyproject.toml
[project]
name = "mastering-mvp"
version = "0.1.0"
description = "Auto-mastering MVP (Next.js + FastAPI + RabbitMQ + Mongo + S3)"
readme = "README.md"
requires-python = ">=3.11"
# Базовые зависимости (общие для api/worker — можно оставить пусто и использовать groups)
dependencies = []

[tool.uv]
# Включаем .venv в корне, чтобы VS Code было проще
managed = true
cache-dir = ".uv-cache"

[project.optional-dependencies]
api = [
  "fastapi==0.114.2",
  "starlette>=0.37.2",
  "uvicorn[standard]==0.30.6",
  "aio-pika==9.4.2",
  "aiormq==6.8.0",
  # DB: Postgres + SQLAlchemy (async) + Alembic
  "SQLAlchemy[asyncio]==2.0.35",
  "psycopg[binary,pool]==3.2.3",
  "alembic==1.13.3",
  "boto3==1.35.19",
  "pydantic==2.9.2",
  "pydantic-core>=2.23.4,<3",
  "pydantic-settings==2.5.2",
  "python-dotenv==1.0.1",
  "click>=8.1.7",
  "h11>=0.14",
  "watchfiles>=0.22",
  "typing_extensions>=4.8",
  "PyJWT[crypto]==2.9.0",
]

worker = [
  "aio-pika==9.4.2",
  "aiormq==6.8.0",
  "boto3==1.35.19",
  "pydantic==2.9.2",
  "pydantic-core>=2.23.4,<3",
  "pydantic-settings==2.5.2",
  "python-dotenv==1.0.1",
  "typing_extensions>=4.8",
]

dev = [
  "ruff>=0.6.9",
  "pytest>=8.3.3",
  "httpx>=0.27.2",
]
